sap-client shall be configured with quotes
If the sap-client number starts with zero (e.g. 001) and it got configured here without quotes, then the proxy will try to connect to sap-client 1 instead of 001.
Furthermore, the proxy log will not point the developer to this wrong sap-client 1. It is quite hard to figure out the wrong configuration.
Hence, we shall give a hint in this documentation, that the client shall be entered always with quotes.