feat(ui5-tooling-modules): extended UI5 modules support (#583)
* feat(ui5-tooling-modules): ignore UI5 modules for transpiling

Enhanced the transpiling script to also follow the dependencies
chain, e.g. if a module is resolve from another NPM package (not
via the UI5 tooling) then the dependencies of this UI5 module needs
to be also covered by the task. For the middleware the client will
follow the dependencies and request them from the server.

* feat(ui5-tooling-modules): parsing XMLView dependencies transitivly

Parses all XMLViews for custom elements and controls and resolves
them from NPM packages if the namespace matches the NPM package
name.

Fixes #578