feat(ui5-tooling-modules): fix AMD bundling and filter Node built-ins
In order to properly support bundling of AMD modules the
rollup-plugin-amd has been replaced with a custom rollup plugin
which uses the latest version of @buxlabs/amd-to-es6. This
ensures a proper bundling of the AMD modules which failed in case
of the tui-image-editor.

To fix the bundling of xlsx, the skip assets plugin has been
extended to allow skipping the loading of Node built-ins which
leads to issues while looking them up with the require.resolve.
This fixes the bundling issues with xslx.

Fixes #623