feat(ui5-tooling-transpile): generate d.ts and filter ts files (#693)
The `ui5-tooling-transpile-task` now also generates the `d.ts` files for the typescript sources. For UI5 libraries, the generation of `d.ts` is active by default and for other project types the generation of `d.ts` can be enabled by setting the configuration option `generateDts` on the task. 

With the generation of the `d.ts` the inclusion of the `ts` sources becomes obsolete. The inclusion of the `ts` sources always was a mistake and has now been removed with this change.