feat(ui5-tooling-transpile): align and clean-up configuration options (#701)
Cleanup of configuration options and properly describe their
meaning and their usage. Explain the usage of external configuration
for Babel and the browserslist.

Defaults the @babel/preset-env to use the "defaults" setting from
browserslist which equals ">0.5%, last 2 versions, Firefox ESR, not dead"
as the target to transpile the classes to.

Renamed the transpileAsync to transformAsyncToPromises to clarify
that this transpiles async/await to Promises instead of using the
regenerator runtime which is the default of Babel. In case of transpiling
async/await with an older browserslist configuration this option is
recommended to be enabled to avoid CSP compliance issues with the
regenerator runtime.