feat: karma-ui5-transpile preprocessor (#711)
With this commit, we showcase the Karma testing for UI5 application
and libraries including code-coverage. Therefore this change adds a
new karma-ui5-transpile preprocessor which can be use to preprocess
the sources of UI5 application and library projects using the
ui5-tooling-transpile tooling extension relying on the common
configuration.

In addition, the ui5-tsapp and ui5-tslib are extended to showcase
the QUnit and OPA testing automation with Karma including code
coverage support for TypeScript sources. The OPA tests are still
JS tests since TypeScript is not yet supported for OPA.