feat(ui5-tooling-transpile): rework to avoid configurations for default usage
The programming language is now derived by the existence of the tsconfig.json
which means the project is a TypeScript project and if no tsconfig.json is
present the tooling extension assumes the project to be of type JavaScript.

The decision of the tooling extension can be overruled by providing config
options in the ui5.yaml. E.g. the TypeScript transpiling can be enabled by
setting the transformTypeScript option to true. Setting it to false will
overrule the existence of tsconfig.json and just transpile JavaScript.

The config option transpileTypeScript renamed to transformTypeScript.

Adopted the tooling extension task and middleware as well as the Karma
preprocessor to use the getConfiguration function to normalize the
configuration.