feat(ui5-tooling-module): support chunks for module bundles
Some dependencies like jsPDF are producing chunks. This commit extends
this tooling extension to support chunks from modules and to include
the chunks also in the build output as well as making it available in
the development server.

As a side effect, the tooling extension now also supports the dynamic
imports transformed by babel-plugin-transform-modules-ui5 for the
rewriting and move into namespace.