feat(ui5-task-zipper): allow to specify the deps to include
The task has been extended to allow to specify the list of UI5
dependencies to be included in the final zip analog to the builder
settings includeDependency. The configuration option includeDependencies
now also accepts an array of strings providing the UI5 dependency name
which is the name specifed in the ui5.yaml > metadata > name property.

```yaml
builder:
  customTasks:
    - name: ui5-task-zipper
      afterTask: generateVersionInfo
      configuration:
        archiveName: simpleapp
        includeDependencies:
          - sap.ui.layout
```

Fixes #550