feat(ui5-tooling-modules): include assets of npm packages
This feature enables the ui5-tooling-modules tooling extension
to include assets from npm packages. The assets to be included
can be either just the npm package name including all assets or
a list of glob patterns to select the assets to be included.
The assets will be put into the proper folder like the required
dependencies. In case of including JS files, those files will
not be processed by the ui5-tooling-modules just copied.

```yaml
builder:
  customTasks:
    - name: ui5-tooling-modules-task
      afterTask: ui5-tooling-transpile-task
      configuration:
        includeAssets:
          "@octokit/core":
            - "dist-web/*.map"
          "tui-image-editor":
            - "dist/**/*.css"
            - "dist/**/*.svg"
```

Fixes #590