fix(ui5-tooling-modules): fallback mode to retry bundling with cjs modules (#729)
On Windows it can be that the build with ES modules failed since the default
exports of CJS modules aren't detected properly. In this case, the tooling
extension falls back to use CJS modules for bundling first which finally works
to create the bundle. Unfortunately, there is no better solution but so far
this happened only in the context of jspdf in which canvg requires raf and the
default export provided with module.exports isn't detected properly on Windows.

Fixes #726