fix(karma-ui5-transpile): also instrument in Babel transformation (#731)
The karma-coverage preprocessor instruments the final content of the
Babel transformation done by the karma-ui5-transpile preprocessor.
This leads to wrong code coverage results since the UI5 AMD-like
module is instrumented rather than the JS code without TypeScript.
Therefore the karma-ui5-transpile preprocessor injects the istanbul
plugin into the Babel transformation that it instruments at the
right point in time.

With this change the "coverage" preprocessor must not be used
in combination with the "ui5-transpile". Only the "coverage" reporter
can be used to create the reports finally.