feat(ui5-tooling-transpile): resolve presets/plugins for babel config options
Ensure to resolve all presets and plugins in the babel configuration options
before running the transform functions - which is necessary for pnpm monorepo
scenarios or when running a build or serve with the globally installed UI5
tooling.

fix(karma-ui5-transpile): use new preprocess step for babel config

To check the existence of the istanbul plugin in the babel configuration
the plugin hooks into the new pre-process phase of the createBabelConfig
function from ui5-tooling-transpile and now is able to handle the babel
plugin config entries which can be string, an array or a config item.

Fixes #810