fix(ui5-middleware-cfdestination): support UAA flow
To enable support for the UAA flow, the option `xfwd` of the
`http-proxy-middleware` is necessary to forward the redirect
host information for the SAML auth flow.

In the `xs-security.json` you need to configure the UI5 server
URL:

```json
{
  "xsappname": "sample-ui5-uaa-uaa",
  [...],
  "oauth2-configuration": {
    "redirect-uris": [
      "http://localhost:8080/login/callback",
      "http://localhost:5000/login/callback"
    ]
  }
}
```

next to the approuter server URL, so that it gets accepted
by the SAML auth flow as a callback.

The change also makes the disabling of the welcome file
configurable to allow the usage of the redirect handling
of the approuter.

The `ui5-middleware-index` has been extended to play properly
together with the `ui5-middleware-cfdestination` so that the
handling of the welcomeFile requests are detected properly and
redirected and handled by the cfdestination middleware.

To support authentication of HTML pages not served by the
approuter coming from the local UI5 server, you can make use
of the `appendAuthRoute` configuration option in the `ui5.yaml`
for the `ui5-middleware-cfdestination`.

Fixes #817