fix(ui5-middleware-simpleproxy): support for express/connect servers (#829)
This change ensures the compatibility with express/connect servers
by using the middlware util from the ui5 tooling to determine the
request path.

Support opening of URLs in the browser causing ERR_CONTENT_DECODING_FAILED
by using the response-interceptor of the http-proxy-middleware to
decode the response. The issue didn't appear for XMLHttpRequests.

Enhanced the demo application to proxy the Tripin service via the
ui5-middleware-simpleproxy.