fix: consistent request path handling of all middlewares (#833)
This fix aligns the request path handling of all middlewares which
rely on the request url to retrieve the proper resource. In case of
loading resources via express.Router the `req.url`, `req.baseUrl`,
and `req.originalUrl` contains all relevant information to know the
context path and the local path. The `baseUrl` and the `originalUrl`
must not be overwritten by middlewares to be able to identify the
source context path of the request.

This cleanup is a precondition to ge the nested routing for the
combination of `cds-plugin-ui5`, `ui5-middleware-cfdestination`,
`approuter` and an underlying UAA.

Related to discussion of #817