feat(cds-plugin-ui5): support for configuration options (#836)
The plugin can be configured using the `cds-plugin-ui5` section in
the `package.json`. To define a custom configFile for any module
you can add a configuration object for each module in the `modules`
section:

```json
{
  "name": "cds-bookshop",
  [...]
  "cds": {
    [...]
    "cds-plugin-ui5": {
      "modules": {
        "ui5-bookshop": {
          "configFile": "ui5.yaml",
          "workspaceConfigFile": "ui5-workspace.yaml",
          "workspaceName": "default",
          "versionOverride": "1.117.0"
        }
      }
    }
  }
}
```

The key is the `moduleId` which is either the directory name for local
apps in the `app` directory of the CAP server or for dependencies the
`name` from the `package.json`. As value you can specify an alternative
`configFile`, `workspaceConfigFile`, `workspaceName`, and/or `versionOverride`.
The location of the config files are relative to the modules' base path.

A second option to override this configuration is the usage of the environment
variable `CDS_PLUGIN_UI5_MODULES`. It contains the JSON string from the
configuration above, e.g.:

```sh
CDS_PLUGIN_UI5_MODULES="{ \"ui5-bookshop\": { \"configFile\": \"ui5-dist.yaml\" } }" cds-serve
```

Fixes #830

---

feat(cds-plugin-ui5): support browsing in UI5 server file listing

Rewrites the URls of the links in the UI5 server file listing to
support browsing when running inside nested routes.