feat(ui5-tooling-stringreplace): allow combination with other middlewares
The stringreplace middleware must not be the last middleware in the
chain and therefore should only intercept the response rather than
sending the response. This allows other middlewares such as the
transpiling to run afterwards.

The middleware must be executed before any other middleware which
finally sends the response (like the transpile or modules middleware).
When called, it hooks into the functions of the server response and
collects the content to finally rewrite it. The final send will be
left to the UI5 tooling development server or other middlewares.

Fixes #665