fix(ui5-tooling-modules): chunks are stored in module name folder (#922)
To simplify the handling of chunks of npm packages, they are now stored
in a folder with the name of the module. With this option a single
`resourceroots` mapping is sufficient to delegate all requests to the
npm package and its chunks to the proper location. This is especially
useful for CDN scenarios when the npm package is not included.

Example: CDN usage + `jspdf`

```html
<html>
  <head>
    [...]
    <script ...
      data-sap-ui-resourceroots='{
        ...
        "jspdf": "resources/jspdf"
      }'
```

Using this `resourceroots` mapping, all requests to `jspdf` and its
chunks are now resolved to `resources/jspdf*`.

* `jspdf` => `resources/jspdf.js`
* `jspdf/0c1H2u3N4k5` => `resources/jspdf/0c1H2u3N4k5.js`

Fixes #915