feat(ui5-tooling-modules): introduce bundle generation
With this feature, the tooling extension now generates a complete
bundle based on the depedencies found in the project and project
dependency sources. It scans for imports in the TypeScript sources,
requires in the JavaScript sources, and used modules in the XMLViews.

The "bundledefs" are now obsolete and can be removed from your projects.
As the bundle generation now scans for all used dependencies and uses
all of them to pass them as entry points to the rollup bundling process
reused dependencies across multiple entry points will finally also be
reused in the final bundle.

Since the bundling process doesn't listen anymore to the request of
individual files during development (middleware use case) we now
watch for changes in all source files and dependencies to identify
the demand to re-create the bundle. This watch mode is enabled by
default and can be disabled on demand in case of performance issues.