fix(ui5-tooling-modules): ignore unkown modules for bundling
When modules have dependencies to unknown or native modules
(e.g. "fs") then the tooling extension tries to include them
but cannot find and open them during the build process which
leads to an empty bundle.

In addition, added a proper error logging in case of bundling
errors to see the stack trace for the build and the middleware.
The build will stop but the middleware just prints the issue
without stopping the server.

Fixes #977