feat(ui5-tooling-modules): improved handling of chunks (#1009)
In the previous version the chunks have been sorted into subfolders
which are identified as first module importing the chunks. This
was done as a follow up of #915 which causes 404s when loading
chunks in CDN scenarios. By putting the chunks into the folders
with the name of the module requiring the chunk the module could
be loaded successfully again when using resourceroots for the
module. (implemented in 3.5.x version)

With the enhancement to load the chunks with relative path names
this feature is not necessary anymore and becomes optional so that
it can be re-enabled by providing the config options chunksPath: true.

In addition, the chunks are put by default flat into the resources folder.
By providing the config option chunksPath: "my/chunk/path" you can provide
a subfolder into which the chunks will be clustered into. When using the
config option addToNamespace: true the chunksPath will be appended to
the %namespace%/thirdparty path so that it is relative to the root
where the modules are put into.