feat(ui5-tooling-modules): improved dependency resolution (#1039)
The chunks needed by third-party packages are resolved against the
root namespace (/resources) which leads to issues in case of running
applications against CDN. Therefore, the chunks are now prepending
the namespace of the application to ensure to resolve relative
against the application.

For some third-party packages like punycode built-in modules in node
exist. In this case, the dependency resolution looked up the node
built-in module and the node polyfills have been used instead of the
project dependency. This has been changed that for entry points
(dependencies who have been declared in any module) are always
first resolved against the local dependencies before the polyfills
are used.

Fixes #1032