feat(ui5-tooling-modules): support including Assets for Web Components
To support all languages and themes, the Seamless Web Components support
need to include a reference to the respective Assets file of the dedicated
library.

The built-in solution using the optional flag includeAssets for the
pluginOptions of the webcomponents plugin can include the assets module
but this module is loaded too late and in some cases the custom theme
and the language isn't available early enough. This requires a rethink
of the OpenUI5Support module.

Manually the Assets can be included as shown in the Components.js of the
sample project:

import "@ui5/webcomponents/dist/Assets";
import "@ui5/webcomponents-fiori/dist/Assets";