fix(ui5-tooling-modules): ensure full rebuild after change of modules (#1177)
The mechanism in the rollup plugin to generate the package wrappers relies
on the lookup of the WebComponentsRegistry. When the packages and classes
are known to the registry the packages will not be emitted anymore. In any
case we nevertheless need to rebuild the web components metadata in case
of changes to them. So the clear function now triggers a complete reload.