fix(ui5-tooling-modules): fixed precendence of module resolution via package.json
The exports field has a higher precedence than the browser/module/main fields in
the root of the package.json. Therefore, we need to change the order of the
resolution to ensure that the exports are handled first.

Another speciality is the handling of the import/require/default fields in the
exports which is import/default/require for all modules and import/require/default
for browser modules.

Specialities are necessary because of the following modules:

  - Firebase will export no modules when using the "default" field instead of the
    "require" field.
  - Axios will not work, when not considering a different module resolution for
    browser modules as it maintains the browser modules.
  - sinon requires to always resolve the exports fields before the root modules

Fixes #1196