feat(ui5-tooling-modules): support HTMLElement as base class for webcomponents
The rollup plugin for webcomponents now properly handles sap/ui/core/html/HTMLElement as a base
class. This allows webcomponents to extend from it and be properly integrated into the UI5
rendering.

The WebComponentRegistry is now able to resolve the superclass for webcomponents in the @ui5/html
namespace to sap.ui.core.html.HTMLElement.

Helper functions isUi5CoreHTMLElement and isHTMLElementBase have been added to
WebComponentRegistryHelper.js to facilitate this change.