fix(ui5-tooling-modules): support npm packages with .js ext (#1254)
An NPM package having the same name like the index file in the
package (e.g. easytimer.js) lead to issues while discovering the
entry modules as the file extension was stripped from the module
name and this leads to a wrong resolution. This change fixes this
issue by checking the module name similarity as well.

Fixes #1251